fix to build scripts for copying plugins so that plists aren't corrupted
-- Mike, this gives me a build that launches and creates a document. 

Can you work on tracking down and eliminating, and I mean eliminating, *all* warnings in the Complete Build? I think that will force the final issues and then we can go from there. (Back to making plugins be elements, etc.)

Note that the toolbar localization is messed up? Is this what you were referring to? I'll work on tracking that down in the morning.