Code Injection Syrynge graphic.
In several sizes so it will look good with UI scaling, and so it will look good at several base scales.

I would recommend 14x14 for the Small site outline size, 20x20 for the Large site outline size.  

Also, I think that the UI would look good to have a larger image in the title area of the code injection window.  Maybe the 40x40 size would be appropriate.