Something I whipped up before I forgot how to make WebKit plugins, as I learned while working on PeaceKeeper.
Just sort of a prototype, with placeholder graphics and NO ACTUAL FUNCTIONALITY! 

The idea is a bit of editing control over image elements.  (Image page/pagelet only for now).  The thought would be the ability to resize, rotate, crop, perhaps add rounded rectangle corners on images using some on-screen  controls that could fade into existence when you move your mouse over an image.

Give this a try just to see what you see.  We can discuss it later; it will be just a few lines of code to disable this for now, since it is a bit of new functionality.  (We could always do baby steps with this.)