Fixed copy of non-existant webkit plugins for now
Logging Image replacement stuff as DJW for now

for DEBUG, setting default to abort when an assertion fails.  At this stage of the game, an assertion is pretty serious and needs to be fixed.

There's an assertion that fails in the HTML parser where its document is not set.  This breaks the possibility of graphical text.