3 new 'minimalist' designs from Sven.
They really need names, though.  They are currently called 'A2', 'A3', and 'Stars'.  They 'Stars' is named after the star motif that appears in various contexts, but there's also the light diagonal backgrounds.

I'm looking for ideas!  A3 sort of reminds me of 1950's bachelor pads with exotic cocktails and lounge music.   A2 is kind of sturdy, industrial -- the orange at the bottom is a bit of an accent.  The orange reminds me a bit of a clownfish/anemonefish, but only slightly... :-)

Any ideas?  It doesn't actually have to be meaningful, just interesting and maybe reminding you of something...