first cut at release notes for this new version.
We'd need to convert to HTML for sparkle, and also merge in the
old release notes of whatever the last item from the 1.2.1 branch is...

SVN release # is included in there just for now, as reference.


This definitely does NOT sufficiently reflect a lot of the fixes that have
been done to the trunk!  It's just that I don't really know how best to 
express what the big differences are from the USER's perspective.

By all means, please add in whatever you think sufficiently conveys that.



Also, added a few new words & took out 'talbot' in the reg list.