Fix to this:
foreachWithParameters:... Caught NSUnknownKeyException: [<KTDocument 0x161ba1a0> valueForUndefinedKey:]: this class is not key value coding-compliant for the key registrationHash


"registratationHash" is one of those methods were were obfuscating for making it harder to crack -- except that we were trying to call that method from a page template, which doesn't know about the #define's.  So I just made the method "hash" and didn't try to obfuscate it.