renamed @interface KTHTMLParser ( Private ) to ( ExtraPrivate ) so that it doesn't conflict with the declaration in KTHTMLParser+Private.h
Note that in checkin 6399 Dan changed all occurrences of ( ) to ( Private ) in category names. 

Mike: apparently a category of ( ) isn't allowed in Tiger so we need to give all categories names. KTHTMLParser now seems to have a lot of moving parts and separate private interfaces. Can you look through and regroup logically?