1) changed SANDVOX_VARIANT TO PRODUCT_VARIANT and added to SUFeedURL in Info.plist so we can check beta or release
2) admitting the error of my ways and shortening CFBundleShortVersionString to just SANDVOX_VERSION, otherwise the Sparkle Update panel just doesn't look right, and you know, I'm all about the aethestics

I was initially resistant thinking this would not let people examine the build number in Finder, since CFBundleShortVersionString is used by Finder to display the version number. But, this appears to be OK, since we also define CFBundleGetInfoString this is used by Finder's Get Info panel and will show our build number (and our copyright), so I think everything is happy now...

Dan: this means that the FB reports probably need to be switched back to read correctly. Maybe we can sync up on this just before we put out the new alpha?

3) Updated SANDVOX_VERSION to 1.5a5 in prep for new alpha