Give the user a chance to get an update to an expired version before they are forced to quit.  This will allow somebody launching an old beta to get the newest version without having to wonder how to get it.
If the app has expired, they are first presented with an alert telling them it has expired, with a button to launch a Sparkle update. They then will hopefully install the new update.

Five minutes later (should be enough time to do this), if Sandvox is still running, it will present the original alert with a "Quit" button.