BUGSID: 28554
new implemetation of autosaveDocument: that uses saveDocumentWithDelegate:::

this is very nice from the NSDocument side of things and seems to work, the remaining problem is that *we* still don't get an error back, so if there was a problem and the document couldn't be saved, we just turn autosave back on again and force the user to resave an unsaveable document. so this might not quite be it...

also, changed

NSString *kKTDocumentType = @"Sandvox Document";

from @"KTDocument" so that we have a constant that matches its use as CFBundleType/UTTypeDescription

also, changed

NSString *kKTDocumentUTI = @"com.karelia.sandvox.site-document";
NSString *kKTDocumentUTI_ORIGINAL = @"com.karelia.sandvox.document";

so that we can separate document types. if you can think of a better name for the type that means "compatible with 1.0-1.2 only", please change