BUGSID:29455 Revamped plugin delegate system to handle undo/redo or deletions.
Separate plugin delegates manager has been scrapped. Instead, the delegate is created on-demand, and the two retain each other, creating a retain-cycle, rather like Core Data properties. When the [KTAbstractElement -didTurnIntoFault] is called, the element releases its delegate, breaking the retain cycle. This means that delegates are dealloced on doc close, or element deletion, and that undoing a deletion instantiates a new delegate for the element.