changed all calls to uploadData:toFile to be uploadFile:toFile: by writing each NSData object to /tmp/<UUID>
preliminary tests suggest publishing is much happier with this

but this is just test code -- it leaves behind a bunch of orphaned files in /tmp

how should we handle this? 

1) write to Application Support? delete dir when publishing finishes?

2) write to .svxSite document? 

there's not much advantage to #2 at this point, perhaps later if we can "stale" things on disk. would speed up the generation phase, perhaps.

would like to discuss options...

other ideas?

so this is the first time we're actually writing everything to disk first and it seems to be working fineL what's the best way to handle local file management?