Made Photo plugin be always just 'Photo' for untitled -- it was inconsistent with Pages and pagelets, but the localization was only for the generic Plugin title.  So this makes everything consistent, and it also fixes the problem of replacing photo page names when you drag in a replacement photo.
Fixed issue where HTML inspector was loading when changing pages in the site outline because we wanted to close it.  Made a new method 'HTMLInspectorControllerWithoutLoading' to just return the HTML inspector if it had already been loaded, nil otherwise, so we didn't load up the inspector just to close it.

Add a method to log javascript errors, so it's easier to know what is happening if we have any javascript errors in our websites that have javascript in them (ours or users')