Only log JavaScript if the hidden default LogJavaScript is set.  It got annoying very fast.  Put this into debug menu though binding to 'value' doesn't actually work!  No biggie though.
Fix 34097.  Saving doesn't convert a URL request to load /dev/null when saving; it just turns the URL request into a cache-or-don't load URL request.  Everything seems to work now when we come back to load the URL from the web view.

(To do in a minute: fix the same disallow over in KTDocWebViewController)


Speed up app launch by loading local fonts only when needed -- only when chosen in the design or when it needs them to generate the QC thing, and only if there are actually fonts in the bundle.  This should save about 3 seconds of launch time!