RELEASE: Fix a licensing bug where some 'Regular Edition' licenses had been treated as 'Pro Edition' licenses.  Such licenses are now being correctly intrepreted.
behind the scenes -- We were literally generating Pro licenses since 9/30/2007!  

So I"m using the same hash scheme to check if the registration is in that list, and if so, downgrading to non-pro capabilities.

Alas there is a big huge honking data file but this can all go away when we switch to 2.0.