wrapped every call to setPermissions: in new default (hidden preference) ConnectionSetsPermissions
currently defaults to YES, but I'm starting to wonder...

we've gotten more than a handful of cases where Host Setup (and Publishing) just out and out fail, and they fail when setting permissions -- putting in test code to bypass the calls to setPermissions: lets those sites I've tested run through Host Setup and Publish just fine

Question: should it be YES or NO by default? what are we gaining by setting permissions on everything?

I suspect the answer is that some sites will absolutely require permissions to be set, some sites will absolutely not publish if we do, and the majority of hosts don't care because the default file permissions on the server are just fine for the user's account. So, it seems that this "hidden preference" will need to be promoted to a switch in the application preferences. For our own support sanity, we may need to do this sooner than later. If would be nice if there were some way we could detect the best choice.

let's discuss...