BUGSID:35835
A) Passing a baseURL into WebKit when applicable to allow relative paths for pro users.

B) Forcing the main document URL request to reload ignoring the cache. WebKit will still cache sub-resources if it feels like it, but ensures that edited CSS etc. files are read in again if needed.