Adding a prototype file, a copy of the standard template, which, if a user default is set, will generate some PHP of the twitter content.
The idea is that the fresh content will be visible to Google, to help increase a page's effectiveness -- also visible to JavaScript-less browsers FWIW.

Not totally done yet... a drawback also it that it will slow down page load since it has to grab the info from twitter, which may be slow.  I think that to be really effective, we'd need some sort of file cache so that multiple page loads don't require a trip over to Twitter.  Without server controls this will turn into a big mess, so maybe we just need to keep this a private thing for us to use, for now...