MERGE: BUGSID:38107 We're getting some mystery exceptions in doc saving doing stuff like:
-[NSConcreteValue UTF8String]: selector not recognized

This is only happening in 1.5.4b1 so I conclude that it has to be a side-effect of the new window rect persistence code. I've tweaked the code to use separate keys for NSString and NSRect versions of the frame so nothing should be able to conflict any more.

Also, move doc window rect restoration code to happen after toolbars are created so that the window doesn't change its size and position slightly.