First cut at adding meta description to UI for 1.6.
Something is wrong with my dependent key declaration in bindings -- not sure what is going on.  Any ideas?

In KTAbstractPage I do this in +initialize:

	[self setKey:@"metaDescription" triggersChangeNotificationsForDependentKey:@"metaDescriptionWarningIsVisible"];
	[self setKey:@"metaDescription" triggersChangeNotificationsForDependentKey:@"metaDescriptionCharCountColor"];





Also can I get your guys' UI ideas?  My idea is to have a progressively stronger warning as you get to the maximum character length of 156.  Right now when you hit 147, as you keep typing the character count gets redder, and if you go above 156, it's all red, and also I put in a warning icon button (which can actualy be clicked to open a to-be-written help page.)

I am thinking that I don't want to mix colors (red and yellow) -- maybe I should have a more and more yellow background, and when you hit 157 the black count turns into red text to say you have gone too far?

Should I also have a character count *down* as well? I don't want to make the UI too complicated....

THOUGHTS PLEASE!  (This will be a valuable addition to 1.6 ... for our users and for us as well on our own website!)