adds support for JS-Kit supplied comments
I tried to do this with a minimum of localization changes. Need to copy UI elements from en's Info.nib to other localizations. We can probably get away with not updating strings for 1.5.4.

Waiting to hear back from JS-Kit on permalink issue, but I figured I might as well check this in.

Probably also should add something to bottom of KTCommentsTemplate.html that displays a message if allowComments but neither commenting system has been activated, but this would add a new string for localization...

Thoughts on UI? On adding new string?

(We can do Disqus comments to, but require more UI and strings changes. Should I add them to 1.6?)