hand-merge of idea to use NSCollectionView for Design Chooser so Mike can take a look at it
to test:

make a new site
add a text page
flip back to home page, so something draws with a design (and outline has a selection)
click Designs button on left side of toolbar

you can select and choose a new design, but it currently won't display the newly selected design until you select a different page

Mike: what's the best way to tell the webview to reload with the new design? the SVWeb* classes don't (yet?) observe the kKTDesignChangedNotification

truth be told, I'm not particularly happy with how NSCollectionView is behaving for this sort of chooser. obviously we can tweak the look -- and categories have yet to be added in this merge -- but do we like the feel of the core animation bits?