Getting close on the bindings of the pge details
when multiple pages selected and extensions are not the same, there should be a placeholder in the popup menu.  Supposedly supported in the nib (but for null placeholder) -- maybe mess with bindings manually?

When multiple pages are selected, we need to lay out the fields according to the placeholder's width.

I'm trying to get the "go" button to hide when there is null/multiple selections.  No luck with that yet....