Figured out a way to get the Web Editor loading behaviour I want using only a single view.
We now use a private WebView delegate method to be informed once WebKit has performed its first layout of the main frame. At that point, the next display update will be to render the content, whether it be mid-load or finished. So, if we jump in just before that next drawing update and check to see if the view has finished loading, we know whether to switch to the progress indicator.

Before, this was accomplished by swapping between 2 WebViews using a smaller timed delay. This new technique increases efficiency (far fewer resources to manage) and simplifies design. We could re-instate a timer-based solution on top as well if needed.