Think I may have had a rather ingenious idea. Adding in a tree of PublishingRecord objects. The idea will be that they maintain a tree representation of everything we've ever published. Most importantly, this will include cached information when known. For now: digest, date, and length.
For example:
When a page is asked if it needs publishing, it can generate a hash, compare that against the corresponding publishing record, and make a decision from there.

Eventually, when publishing Sandvox should be able to finish testing the contents of the site and produce a list of records which can now be deleted, asking if you wish it to be so.