Added Sven's ESCursor
Mike, I may have messed up the relative paths for "Shared in the project file.  If so, can you try to fix on your end?  It works here but I know when I add shared code, it's hard to get it right.

Also: fixed file's owner for doc inspector nib.