Stopped using SVBlogSummaryDOMController for now
In web editor view, set the appropriate angled cursor based on which handle is clicked

When resizing, snap to the original image's dimensions when you are within just a few pixels of it.  Also, don't let the image get less than 16 pixels in size since you can't really resize it if it gets too small.