Fixing a bunch of methods that create NSError messages.  In many cases we were creating strings that we wanted to show as secondary information below the main error message.  Well, it turns out that NSLocalizedRecoverySuggestionErrorKey is the way to go with this ... as the docs say, "This string is suitable for displaying as the secondary message in an alert panel."
I'm really trying to get the error message presented upon failure to open a document to be a lot friendlier.  Still a couple of adjustments I want to do, like wrapping an NSError in a friendlier message.


Also, standardizing on NSErrorFailingURLStringKey as a constant, which is defined in 10.5 SDK -- though it is deprecated for 10.6 SDK, it's just the constant name; no big deal.