observe the selected webview controller so that we change bindings only when appropriate.
Clear the title & meta description from the controller when we start a load, or go out of visibility, so we aren't holding onto old values.

Try to get the meta description when the <title> tag is received.  Alas we don't have a callback for when the <head> is loaded, so this only works if the meta description came before the <title> in the source.  So we also check after the body has fully loaded, but that is often a bit of a delay.