Simplified KTLinkConnector greatly to not require clients to have all those delegate methods. Drag pasteboard is now encapsulated.
The only trick I'm not sure how to approach -- TO DISCUSS WITH MIKE -- is how to best encode the target page on the drag pasteboard, so that when the drag is done, we can see what page we dragged to.  We were using a unique ID and then asking the MOC to give us the page object, but it's not clear how to get a MOC from a view (non-hackishly).  Can we just put a numeric pointer to the actual object on the pasteboard, since it's only within the application?