initial implementation of drag-and-drop using NSPasteboardReading
unclear that KSWebLocation is handling all the types that the original plugin did

need to discuss testing plan

can we set title via URL? original code suggests we can, for some types. if so, how would do that that using initWithPasteboardPropertyList since we don't yet have a container to set the title on?