Trying to make nice effects on image thumbnails in site outline kind of like the old version, but via code rather than bitmap masks now that we're not using page curls.
processForThumbnailOfSize is my attempt to do this in core image.  However I'm not really that happy with things -- The cropping is not working properly, so we end up with an incorrectly scaled image.  Also, the code seems to be called repeatedly; the image should probably be cached.

Can we code review this?