Loading HTML into the HTML editor now.  Not saving back yet or getting window title.
Validating the toolbar & menu actions for raw HTML editing.

We're validating menu and toolbar items a bit more vigorously now.  You may find that your favorite menu or toolbar items are disabled -- if so, please check the  validation code and put in whatever code you need to validate whether it's allowed to be performed.