#81837 collection empty preset now has its icon.  I'm now using the empty page icon, but I wonder if we want it to look even *more* empty to sort of align with the other presets, whose icons only indicate the  indexes themselves.
I'm thinking we will want to re-do all the index page icons anyhow ... maybe it makes sense to have two variations -- one for the index itself (maybe something similar, but in a shorter rectangle indicating it's an object on a page, not a whole page)  and then one that is similar to what we have here, but looking more like a page.

An alternate would be to really embrace the "folder" metaphor, and superimpose what looks like an index over a standard folder icon.   So it really does look like a collection, with its contents indexed on the outside.  Hmmm...