#81577
added YRKSpinningProgressIndicator to project

NB: this is now a separate (forked) repository on github.com/karelia. You'll need to 

git clone git@github.com:karelia/yrk-spinning-progress-indicator.git

in the same directory you have KTHTMLWriter et al.

I'm on the fence as to whether this is the right approach to importing a single third party class, but it is theoretically cool that we could check in our own mods, so let's see how this goes in practice.