adding in specific Update* targets for each Karelia repository on GitHub
these all needed to have 

DEPLOYMENT_POSTPROCESSING = YES

added to their Release configuraitons, with configurations based on Sandvox, or they weren't actually being executed at build time

since we own the repositories -- at least these particular forks--, we're going to assume we need to pull ToT for each build. I can't really figure a way to request a certain revision number, like we do for iMedia, so we're going to assume ToT for now.

also added syncing iMedia with IMEDIA_REVISION in Sandvox.xcconfig back in so that we're staying up to date with iMedia 2 progress in these builds. Should we do away with a specific build number and just go with ToT?