Hook up choose button to do the right thing.
Also added -- not hooked up yet -- a popup to match the look/feel of other image inspectors, so you can have no poster frame, an automatically generated one, and a custom one.  I guess I'll want to hide the choose button and image well unless the custom is chosen.

(I guess I am going to have to have a property for which one ... otherwise there is no real way to know if I had an automatic or custom one set if I change the source video...