I think this works pretty well.  Graphics to insert have a subtitle with a description.
Some of these descriptions come either from SVGraphicFactory graphicDescription -- we need to come up with some better text so if you have any ideas please make the changes.

The rest come from the plugins themselves.  We have had a key KTPluginDescription that has been unused since day one -- so that's a good place to start.  I'd like to make some of them be a bit more terse.  We do need to get these into the localized strings files for the bundles; I'll start on that part.