Experiment for #95978. clear:both on all callout and graphic containers *except* if it's the first element in text. (Could be done with a CSS pseudo-selector if we drop IE6 support)
This means no special clear rule for existing sites (on the whole), but does introduce the clear behaviour for sites that take advantage of positioning a graphic customly elsewhere in the doc.

Some designs don't float their main content, so this then pushes the graphic below the sidebar. 11Mystics & Jumsoft designs are all fine, but about half of Charlies exhibit this behaviour. About 1/3 of ours do too. I think it could be reasonable to expect designs to be updated to accommodate this (override to do clear:left instead e.g.)