friendlier prompt.  Need to make sure it does the right thing.
Occasionally, but not always, I lose the selection during drag.  I think that the live updating with bindings is causing the issue.  We may need to bind to an inspector, and then push the final value as I did in the general index inspector.  Or, take out the live updating, but then you can't really see what you are doing until it's over.