add French localization for testing
which of course immediately pointed up the problem with our solution:

in English you might say "2 hours ago", but in French you would say "il y a 2 heures"
which blows away the possibility of a simple, in-order string substitution...

I'd rather not have to write a separate non-localized localizing time function for each language, but what else can we do?

To be clear: this isn't something Twitter provides, but a clever hack we're adding on. Do we really need a timestamp?