Since Colorbox functinality is now visible in Photo Grid Index, we might as well start bundling the gallery index as well.
RELEASE:  New "Gallery" functionality, accessible via its own new index type, or as an option on a photo grid index, to present images in a collection. 

KNOWN ISSUES - Changing an attribute of a gallery causes it to disappear from the web view; you need to reload the page to get it again.