use writePlaceholderWithText:: to write out placeholder
now to find a way to shorten it. I suspect that "Please specify an address for the recipient using the Inspector." has confused many people over the years. In most cases, YOU are the recipient, but not necessarily. You could be setting up a site for somebody else or want to use info@ or whatever. How to communicate that in a clearer, shorter way?

Mike?