see Case 82975 -- I don't think meaningful Digg migration into an RSS feed is really achievable. RSS is now a very limited output mechanism for Digg (like 8 feeds total). *A* possibility would be to support the old plugin, at least for migration. We just need to repair our script such that it parses and does DOM writing correctly using the jQuery that we're shipping in S2.
So, don't panic, I just whipped this up quickly as a point of discussion. The plugin is converted, with the exception of fixing our diggsandvox script and localization.

The other possibility, now that I understand how to build plugIns from when I first looked at Digg, is that I could probably whip up something that is a subset of Digg's new widget generator and migrate S1 info into that. 

http://about.digg.com/downloads/widgets

I'm not sure how to cram all of those options into the Inspector (maybe tabs?). I'm also not sure how to handle color selection in the Inspector. Maybe we do a subset of options? I actually think the new widget code from Digg is much more compact and fits in with the S2 API quite nicely. I think it's a days work or less modulo other issues.

So, checking this in -- though not yet integrated into S2 build -- as a placeholder. 

I actually think we should ship something in the way of a Digg for S2, whether with the old look or the new. (I favor the new if we can tame it.)

May I suggest a Skype?