#121430 - RELEASE: fix issue with loading General Index inspector under Leopard
so it turns out that the Layout popup button's menu had an "Accessibility Title" set on each item. when AppKit hits loadView it throws an exception, our setInspector: code catches the exception, bypassing the rest of the loading of the secondary nib. not good.

2011-05-30 13:59:57.207 Sandvox[15262:813] *** -[NSMenuItem accessibilityIsIgnored]: unrecognized selector sent to instance 0x16733230

it would appear that having menu items which are images only, as in our Layout popup, and have accessibility info set, really only works under 10.6:

We added support for -accessibilitySetOverrideValue:forAttribute: to NSMenu and 
NSMenuItem, for simple value types in 10.6 (NSString, NSNumber), primarily to 
enable adding an accessibility description to menu items which display only an 
image. This also included implementing accessibilityIsIgnored on those two 
classes in 10.6.

-James Dempsey


only solution I've found so far is simply to remove those descriptions from the xib using 3.2.5 IB so that's what I'm doing.