Logging the data size of the key in a dictionary.  Trying to diagnose the following memory overflow message:
malloc: *** mmap(size=15859712) failed (error code=12)
*** error: can't allocate region
*** set a breakpoint in malloc_error_break to debug

(I'm setting a custom thumbnail for a video, from a TIFF file that's 15.1 megabytes)

My logging shows this:

addMediaWithRequest: about to setObject forKey of data with length of 15859240, 15.1 MB

Maybe a leak?  Or some other problem with using a 15.1 MB NSData as a key in a dictionary?